Starting to add sample configuration and documentation of tooling modules
For now, there's a bunch of explanation via inline comments in this sample module.
This will probably be pulled out into a wiki or readme file later.  It's helpful
to see the sample code, though, when discussing the relationships between nodes,
sites, services and roles.  The design of the DscConfiguration module and the
Resolve-DscConfigurationProperty command didn't really click for me until I saw
a sample configuration script much like this one.

Still to do, build a sample configuration data DB, and update the configuration
sample to use it, rather than hypothetical usages in comments.