Moved Resolve-DscConfigurationProperty to DscConfiguration module
It was likely a part of dscbuild due to how it used to use Get-Variable -Scope ## to look up the value of the $ConfigurationData variable in parent scopes.  This was probably intended to find the automatic variable of that name inside the DSC configuration functions, but Get-Variable doesn't cross module scope boundaries.

This code has been updated to use $PSCmdlet.GetVariableValue() instead, which looks up the ConfigurationData variable in the caller's scope stack (regardless of which module the caller happens to be in), allowing the function to reside in the DscConfiguration module, where it makes the most sense.