Updated resolution order, removed DefaultValue parameter
Decided yesterday to remove the DefaultValue parameter, after a chat with Steve Murawski.  This is to force all such settings to be defined in the configuration data files, instead of splitting this knowledge between the data files and the configuration function.  Default values should be defined either in Services, or in the 'All' site (depending on what the value's intended use is.)

Also changed the resolution order so Site settings take precedence over Services.  (Possibly more to come along these lines; we're still brainstorming about the use cases that need to be covered.)