Revert "Get-TargetResource only requires Key properties, not Required."
This reverts commit ec24afd5c92bd1a84bd0d6ae51b8093fc0ab5866.

Turns out the LCM really does require Get-TargetResource to accept all of the Required properties in addition to Key.  The error message indicates that it's a missing "key property", which probably makes this a bug in the LCM of v4.0.

Since this module's purpose is to make sure a resource will work when deployed to a node, we need to make sure the behavior of the LCM is enforced, even if this behavior does turn out to be a bug.