Node to Service relationship update
Nodes can now optionally have a Services key which can have one or more service names (similar to how Services currently contain a Nodes key to define the relationships from the other direction.)  When Resolve-DscConfigurationProperty is called, both places are checked to determine whether a particular service is associated with a particular node.

This makes it easier to programatically generate a new Node file and associate it with services.

A corresponding Services parameter has been added to the New-DscNodeMetadata command.

This commit also removes some leftover code related to the old "Roles" key.