Administrator rights changes
cDscResourceDesigner and DscConfiguration (both dependencies for the DscDevelopment module) were not able to load unless you're running PowerShell "as Administrator."  This is an annoying experience on a developer machine, particularly since the Git client complains when you _are_ running as administrator.

cDscResourceDesigner already contains admin privilege assertions in the couple of commands that require it, so there's no reason to have #requires -RunAsAdministrator for the whole module.

DscConfiguration was supposed to be swallowing the error if you don't have the ability to query the LCM (non-admin, basically), but the error was apparently non-terminating and still causing the import to fail.  Added -ErrorAction Stop so the try/catch will work.

Now I can load up DscDevelopment in a non-admin session and run commands like New-DscResourceShell, etc.