DscBuild Updates
Invoke-DscBuild no longer needs to make changes to the C:\Program Files\WindowsPowerShell\Modules directory when it runs.  Modules are tested directly in the SourceResourceDirectory, and configurations are compiled with the modules in that location as well.

Resource modules may now optionally contain a Deploy.ps1 script, which accepts a single positional parameter that is the directory into which their deployable contents should be copied.  This folder will already exist, and the module should not create its own root subfolder.  This allows modules which contain large files for documentation or testing to leave those out of the zip file that's distributed via the pull server.

If a module does not contain a Deploy.ps1 script, then its entire contents will be packaged and deployed.