Added ModulePath parameter to Invoke-DscBuild
By default, only $SourceResourceDirectory and $pshome\Modules will be on the PSModulePath while Invoke-DscBuild is executing.  The -ModulePath parameter allows callers to specify additional folders that should be on the path during execution as well.  This will typically be set to the path to the folder where $ConfigurationModuleName can be found.