Updated DscConfiguration for ProtectedData 4.1.0+ compatibility
ProtectedData was recently updated to deprecate the -SkipCertificateVerification switch, which we were using.  Right now, it would still work, but would generate warnings.  Eventually, the switch will be removed.

This new version of DscConfiguration just bumps up the requirement so it must be using ProtectedData 4.1.0, and removes the deprecated switch.